Define `impersonate_user_id` claim
This should really be `impersonate_user` but because of a mistake
in the previous definition (maxLen of 32 instead of 36), we cannot
use `impersonate_user` as the name of the claim. Once Cognito has
a claim defined, you cannot change anything about it.

This also updates to the Landlord image that supports this claim
name.