Add build for mkdocs
* Add mkdocs.yml for configuring the mkdocs site
* The docs site is built in the "build" directory. A virtualenv
for the python installation is created. We also use the "awesome-pages"
plugin to auto-generate the docs for the Terraform modules.
* Some formatting fixes in Markdown. Most notably, requiring an empty
line after a ':' before a '*' list. Without the empty line, the mkdocs
markdown interpreter does not interpret that as a list.
* Number the introduction and installation docs to clarify ordering.
* Move docs that are not ready for prime time to a "drafts" folder.
The are not included in the built docs site.
* Remove all trailing spaces from lines to avoid an annoying warning.
* Some other minor touch ups to the docs (they still need a thorough
editor's pass)